passed down day and hour info to each entry card

+++ src/components/Timetable/Headers/Headers.tsx
@@ -7,7 +7,7 @@ import { DAYS_OF_WEEK } from "../../../constants/appConfig";
 const Headers = () => {
   const renderHeadersJSX = () => {
     return DAYS_OF_WEEK.map((day, i) => (
-      <span key={`${day}-${i}-header`}>{day.short}</span>
+      <span key={`${day}-${i}-header`}>{day.long}</span>
     ));
   };
